% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bca_coverage.R
\name{bca_coverage}
\alias{bca_coverage}
\title{Estimate the number of BCa bootstrap replicates}
\usage{
bca_coverage(A, ci = 0.95, perc_dev = 10, prob = 0.95,
  boots = NULL, point = NULL)
}
\arguments{
\item{A}{A numeric value, indicating the acceleration factor. Can be estimated from the data 
(i.e., using jackknife) or, if A is bnull, will be estimated from boots as in coxed::bca, 
but using the point estimate instead of mean of boots in the calcualtion.}

\item{perc_dev}{A numeric in (0, 100), indicating the maximum percentage deviation of each 
of the upper and lower confidence interval lengths. The lower this value, the more boostraps 
are typically required.}

\item{point}{A numeric value. The sample statistic from the original dataset.}

\item{prob:}{A numeric n (0,1), indicating the approximate probability that perc_dev holds.}
}
\value{
An integer, indicating the estimated number of bootstrap replicates. This value is either the 
initial estimate (if boots is null) or the final estimate (if boots is not null).
}
\description{
Estimate the number of BCa bootstrap replicates
}
\details{
"The accuracy obtained by a given choice of B is random, because the bootstrap simulations are random.
 To determine an appropriate value of B, we specify a bound on the percentage deviation, denoted pdb, 
 and we require that the actual percentage deviation is less than this bound with a specified probability, 
 1 - r, close to one. 
 The three-step method takes pdb and r as given and specifies a data-dependent method 
 of determining a value of B...such that the desired level of accuracy is achieved. For example, one might 
 take (pdb, r) = (10, .05). In this case, the three-step method determines a value B* such that the 
 percentage deviation of the upper and lower confidence interval lengths is less than 10% each with 
 approximate probability .95." (Andrews & Buchinsky, 2002)
Authors recommend iterating through this process for (possibly) better finite accuracy.
}
\examples{

}
\references{
Andrews, D. W., & Buchinsky, M. (2002). On the number of bootstrap repetitions 
for BCa confidence intervals. Econometric Theory, 18(4), 962-984.
}
\author{
Eliezer Pickholtz (eyp3@cornell.edu)
}
