% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_rptgam.R
\name{print.rptgam}
\alias{print.rptgam}
\title{Print an rptgam object}
\usage{
\method{print}{rptgam}(x, digits = 2, ...)
}
\arguments{
\item{x}{An rptgam object.}

\item{digits}{Integer indicating the number of digits to round. Default is 2.}

\item{...}{None used currently.}
}
\value{
The repeatabilities (rpt) (adjusted and unadjusted), LRT-test p-values, permuations p-values (adjusted and unadjusted), AIC values, penalized and unpenalized model p-values
}
\description{
Print an rptgam object
}
\author{
Eliezer Pickholtz (eyp3@cornell.edu)
}
